Add support for permission management and running as bot users
By default, only the owner and designated "sudo" users may use the bot. Commands may specify that other levels of permission can use them using decorators, but these are ignored by default, as there is a bounding set (any permissions outside this set are discarded).
By default, there are no sudo users, not even me, so there is no risk of abuse of the bot.
Additional users can be added through the web interface. The loader exports a function to check the security clearance of a message based on specific flags, so it is possible for a module to provide sub-command security checks.
Changes to permissions are ignored until the bot is restarted, in case someone is able to grant themselves additional clearance through a bug, in which case they still have to wait for a restart. The web interface does not have an option to restart the bot so that is an additional protection. Due to the high levels of trust bestowed on the web interface with this update, it now will send a notification to the owner of the bot to inform them a new login took place, including the IP address.
We also removed the need for a register() function in modules (in fact, its entirely removed)
There is a new translation system to support bots better - it can use context from the message